requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
sqlacodegen
pydantic
python-dotenv
